requirements.txt
psutil==5.7.0
pygount==1.2.3
SQLAlchemy==1.3.19
requests==2.24.0
sentry_sdk==0.19.3
discord.py==1.5.1
praw==7.1.0
base36==0.1.1
discord==1.0.1
GitPython==3.1.7
python-dotenv==0.14.0
PyYAML==5.3.1
psycopg2==2.8.5
Quart==0.13.1
requests-oauthlib==1.3.0
expiringdict==1.2.1
cachetools==4.1.1
